radicle: fix divergence check
The `graph_descendant_of` method is strict in that it will return
false if the two `Oid`s are equal -- they are strictly not
descendants.

If the `head` is the same as the `canonical_oid`, no error should be
returned and the `push` can happen without any issues.

Add a check for `head != canonical_oid` before the
`graph_descendant_check`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>
X-Clacks-Overhead: GNU Terry Pratchett